refactor(ChatPane): extract ChatScrollControls into its own module
Moves the ~307-line ChatScrollControls forwardRef (+ ChatScrollControlsHandle +
USER_SCROLL_INTENT_MS) out of ChatPane.tsx into ChatScrollControls.tsx.
ChatPane.tsx drops 926->597; the trailing helpers stay (none are referenced by
the scroll component). Pure component extraction — exact body, props, ref
handle, and scroll math preserved; the lone Lingui aria-label traveled with the
JSX. ChatPane.test.tsx (41) green.
